refactor: place scene in separate class
