+++ src/Application.cpp
@@ -1,0 +1,8 @@
+//
+//  Application.cpp
+//  Luminosity
+//
+//  Created by Mitchel Wijt on 07/12/2021.
+//
+
+#include "Application.hpp"
