--- src/Shaders/Lighting/FragmentShader.glsl
@@ -1,7 +1,0 @@
-#version 330 core
-out vec4 FragColor;
-
-void main()
-{
-    FragColor = vec4(1.0);
-}
