--- src/Shaders/Lighting/VertexShader.glsl
@@ -1,12 +1,0 @@
-#version 330 core
-
-layout (location = 0) in vec3 aPos;
-
-uniform mat4 modelMatrix;
-uniform mat4 viewMatrix;
-uniform mat4 projectionMatrix;
-
-void main()
-{
-    gl_Position = projectionMatrix * viewMatrix * modelMatrix * vec4(aPos, 1.0);
-}
