+++ src/Shaders/Object/VertexShader.glsl
@@ -11,11 +11,15 @@ out vec3 fragPos;
 uniform mat4 modelMatrix;
 uniform mat4 viewMatrix;
 uniform mat4 projectionMatrix;
+uniform bool hasLightComponent;
 
 void main()
 {
-    texCoord = aTexCoord;
-    normal = mat3(transpose(inverse(modelMatrix))) * aNormal;
+    if(!hasLightComponent) {
+        texCoord = aTexCoord;
+        normal = mat3(transpose(inverse(modelMatrix))) * aNormal;
+        fragPos = vec3(viewMatrix * modelMatrix * vec4(aPos, 1.0));
+    }
+    
     gl_Position = projectionMatrix * viewMatrix * modelMatrix * vec4(aPos, 1.0);
-    fragPos = vec3(viewMatrix * modelMatrix * vec4(aPos, 1.0));
 }
