+++ src/imgui.ini
@@ -9,7 +9,7 @@ Size=298,197
 Collapsed=0
 
 [Window][Sidebar]
-Pos=2,2
+Pos=5,3
 Size=216,655
 Collapsed=0
